Add anchor to slack
Super small change - as of right now hackclub.com/community just redirects to the main page. Hopefully with an anchor it will automatically scroll down the the sign up form. I don't use React much so if this isn't the proper way to do it please lmk and I'll fix it.