fixed a few naming errors
an image was having .JPG as filetype so it wasnt being rendered on vercel coz of linux being case sensitive (not sure if its like that for filetypes but i should test it once).
Also i found a problem in hc site, the thing is that it crashes when we open inspect elements, i use chrome so prob it will be like that on any other chromium based browser, i think its big problem and caused due to hc site being heavy, i will try to look at it but probably its coz of the ancient files this repo has accumulated over a decade.