changes in the hero section
alr so i made some huge changes in the hero section, ealier it was using assemble's group image as the main photo and bg was a lined paper image i stole from scrapyard's website, but the photo from outernet is such a vibe, i asked a few people which of the designs is looking better and they said that the photo is really good, so i put the outernet photo as the main center photo, widened it, and put the assemble photo as the bg rather than the lined paper with a black tint/shade. So rn its looking quite good.